Move corePlugins/index.js to corePlugins.js
